skills/polymarket-trader/requirements.txt
py-clob-client>=0.17.0
requests>=2.28.0
google-generativeai>=0.3.0
